+++ srcs/equations.c
@@ -6,7 +6,7 @@
 /*   By: bede-fre <bede-fre@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/26 12:49:57 by bede-fre          #+#    #+#             */
-/*   Updated: 2018/08/09 12:51:46 by lguiller         ###   ########.fr       */
+/*   Updated: 2018/08/09 15:23:45 by lguiller         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,15 +19,15 @@ static t_rot	ft_rotation(t_scene *tp, t_mat3 *uni, t_mat3 *pos)
 	new.uni.x = uni->x;
 	new.uni.y = uni->y;
 	new.uni.z = uni->z;
-	new.uni = ft_rot_x(new.uni, ft_rad(-(double)tp->dx));
-	new.uni = ft_rot_y(new.uni, ft_rad(-(double)tp->dy));
-	new.uni = ft_rot_z(new.uni, ft_rad(-(double)tp->dz));
+	new.uni = ft_rot_x(new.uni, ft_rad(-tp->dx));
+	new.uni = ft_rot_y(new.uni, ft_rad(-tp->dy));
+	new.uni = ft_rot_z(new.uni, ft_rad(-tp->dz));
 	new.pos.x = pos->x - tp->px;
 	new.pos.y = pos->y - tp->py;
 	new.pos.z = pos->z - tp->pz;
-	new.pos = ft_rot_x(new.pos, ft_rad(-(double)tp->dx));
-	new.pos = ft_rot_y(new.pos, ft_rad(-(double)tp->dy));
-	new.pos = ft_rot_z(new.pos, ft_rad(-(double)tp->dz));
+	new.pos = ft_rot_x(new.pos, ft_rad(-tp->dx));
+	new.pos = ft_rot_y(new.pos, ft_rad(-tp->dy));
+	new.pos = ft_rot_z(new.pos, ft_rad(-tp->dz));
 	new.pos.x += tp->px;
 	new.pos.y += tp->py;
 	new.pos.z += tp->pz;
@@ -39,13 +39,13 @@ double			ft_sphere(t_scene *tp, t_mat3 *uni, t_mat3 *pos)
 	t_equ	p;
 
 	p.a = pow(uni->x, 2.0) + pow(uni->y, 2.0) + pow(uni->z, 2.0);
-	p.b = (uni->x * (pos->x - (double)tp->px) +
-		uni->y * (pos->y - (double)tp->py) +
-		uni->z * (pos->z - (double)tp->pz)) * 2.0;
-	p.c = (pow(pos->x - (double)tp->px, 2.0) +
-		pow(pos->y - (double)tp->py, 2.0) +
-		pow(pos->z - (double)tp->pz, 2.0)) -
-		pow((double)tp->p4, 2.0);
+	p.b = (uni->x * (pos->x - tp->px) +
+		uni->y * (pos->y - tp->py) +
+		uni->z * (pos->z - tp->pz)) * 2.0;
+	p.c = (pow(pos->x - tp->px, 2.0) +
+		pow(pos->y - tp->py, 2.0) +
+		pow(pos->z - tp->pz, 2.0)) -
+		pow(tp->p4, 2.0);
 	p.d = pow(p.b, 2.0) - 4.0 * p.a * p.c;
 	if (p.d < 0.0)
 		return (-1.0);
@@ -66,7 +66,7 @@ double			ft_plane(t_scene *tp, t_mat3 *uni, t_mat3 *pos)
 	t = -(((tp->univect.x * (pos->x - tp->px))
 		+ (tp->univect.y * (pos->y - tp->py))
 		+ (tp->univect.z * (pos->z - tp->pz))
-		+ (double)tp->p4) /
+		+ tp->p4) /
 		((tp->univect.x * uni->x)
 		+ (tp->univect.y * uni->y)
 		+ (tp->univect.z * uni->z)));
@@ -80,15 +80,15 @@ double			ft_cone(t_scene *tp, t_mat3 *uni, t_mat3 *pos)
 
 	new = ft_rotation(tp, uni, pos);
 	p.a = pow(new.uni.x, 2.0) -
-		(pow(new.uni.y, 2.0) * tan(ft_rad((double)tp->p4))) +
+		(pow(new.uni.y, 2.0) * tan(ft_rad(tp->p4))) +
 		pow(new.uni.z, 2.0);
-	p.b = ((new.uni.x * (new.pos.x - (double)tp->px)) -
-		((new.uni.y * (new.pos.y - (double)tp->py)) *
-		tan(ft_rad((double)tp->p4))) +
-		(new.uni.z * (new.pos.z - (double)tp->pz))) * 2.0;
-	p.c = (pow(new.pos.x - (double)tp->px, 2.0)) -
-		((pow(new.pos.y - (double)tp->py, 2.0)) * tan(ft_rad((double)tp->p4))) +
-		(pow(new.pos.z - (double)tp->pz, 2.0));
+	p.b = ((new.uni.x * (new.pos.x - tp->px)) -
+		((new.uni.y * (new.pos.y - tp->py)) *
+		tan(ft_rad(tp->p4))) +
+		(new.uni.z * (new.pos.z - tp->pz))) * 2.0;
+	p.c = (pow(new.pos.x - tp->px, 2.0)) -
+		((pow(new.pos.y - tp->py, 2.0)) * tan(ft_rad(tp->p4))) +
+		(pow(new.pos.z - tp->pz, 2.0));
 	p.d = pow(p.b, 2.0) - 4.0 * p.a * p.c;
 	if (p.d < 0.0)
 		return (-1.0);
@@ -109,11 +109,11 @@ double			ft_cylinder(t_scene *tp, t_mat3 *uni, t_mat3 *pos)
 
 	new = ft_rotation(tp, uni, pos);
 	p.a = pow(new.uni.x, 2.0) + pow(new.uni.z, 2.0);
-	p.b = ((new.uni.x * (new.pos.x - (double)tp->px)) +
-		(new.uni.z * (new.pos.z - (double)tp->pz))) * 2.0;
-	p.c = ((pow(new.pos.x - (double)tp->px, 2.0)) +
-		(pow(new.pos.z - (double)tp->pz, 2.0))) -
-		pow((double)tp->p4, 2.0);
+	p.b = ((new.uni.x * (new.pos.x - tp->px)) +
+		(new.uni.z * (new.pos.z - tp->pz))) * 2.0;
+	p.c = ((pow(new.pos.x - tp->px, 2.0)) +
+		(pow(new.pos.z - tp->pz, 2.0))) -
+		pow(tp->p4, 2.0);
 	p.d = pow(p.b, 2.0) - 4.0 * p.a * p.c;
 	if (p.d < 0.0)
 		return (-1.0);
